Add files via upload
These files contain python code to connect to MySQL and create a database for a fictional electronic appliances store called INIZON. It also contains code that a user can run in order to access the database as a customer. It will give you a list of different types of appliances and further choices of brands and models within those categories. Once you are done selecting it will ask for a few more tidbits, and then generate a bill. I plan to continue working on it and add an admin section to add and remove entries in the database. Currently it is built to run in the IDLE that comes prepackaged with the base python install. I will try and make it run in Visual Studio Code as soon as I figure out the damn thing.